requirements.txt
Flask==2.3.3
pylint==2.17.5
python-dotenv==1.0.0
pyuwsgi==2.0.22
